Rename mw/log/test to mw/log/tests for consistency
