+++ score/mw/log/detail/text_recorder/BUILD
@@ -171,7 +171,7 @@ cc_test(
         "@score_baselibs//score/os/utils/mocklib:path_mock",
         "@score_baselibs//score/mw/log/detail:backend_mock",
         "@score_baselibs//score/mw/log/detail:circular_allocator",
-        #"@score_baselibs//score/mw/log/test/console_logging_environment",
+        #"@score_baselibs//score/mw/log/tests/console_logging_environment",
         "@googletest//:gtest_main",
     ],
 )
@@ -212,7 +212,7 @@ cc_test(
     deps = [
         ":non_blocking_writer",
         "@score_baselibs//score/os/mocklib:unistd_mock",
-        #"@score_baselibs//score/mw/log/test/console_logging_environment",
+        #"@score_baselibs//score/mw/log/tests/console_logging_environment",
         "@googletest//:gtest_main",
     ],
 )
