+++ score/mw/log/tests/BUILD
@@ -20,7 +20,7 @@ cc_unit_test_suites_for_host_and_qnx(
         "@score_baselibs//score/static_reflection_with_serialization/serialization:size_visitor_ut",
         "@score_baselibs//score/static_reflection_with_serialization/serialization:visitor_type_traits_ut",
         "@score_baselibs//score/static_reflection_with_serialization/visitor:visitor_ut",
-        ##"@score_baselibs//score/mw/log/test/log_stream_ara_core_types:log_stream_ara_core_types_ut",
+        ##"@score_baselibs//score/mw/log/tests/log_stream_ara_core_types:log_stream_ara_core_types_ut",
     ],
     visibility = ["@score_baselibs//score/mw/log:__pkg__"],
 )
