+++ score/mw/log/tests/my_custom_lib/BUILD
