+++ score/mw/log/tests/my_custom_lib/my_custom_type_mw_log.cpp
