+++ score/static_reflection_with_serialization/serialization/BUILD
@@ -49,7 +49,7 @@ test_suite(
         ":visitor_type_traits_ut",
     ],
     visibility = [
-        # "@ddad@score_baselibs//score/mw/log/test:__pkg__",
+        # "@ddad@score_baselibs//score/mw/log/tests:__pkg__",
     ],
 )
 
@@ -63,8 +63,8 @@ cc_test(
     features = _compiler_warning_features,
     tags = ["unit"],
     visibility = [
-        #"@score_baselibs//score/mw/log/test:__pkg__",
-        # "@ddad@score_baselibs//score/mw/log/test:__pkg__",
+        #"@score_baselibs//score/mw/log/tests:__pkg__",
+        # "@ddad@score_baselibs//score/mw/log/tests:__pkg__",
     ],
     deps = [
         ":serialization",
@@ -81,8 +81,8 @@ cc_test(
     features = _compiler_warning_features,
     tags = ["unit"],
     visibility = [
-        #"@score_baselibs//score/mw/log/test:__pkg__",
-        # "@ddad@score_baselibs//score/mw/log/test:__pkg__",
+        #"@score_baselibs//score/mw/log/tests:__pkg__",
+        # "@ddad@score_baselibs//score/mw/log/tests:__pkg__",
     ],
     deps = [
         ":serialization",
