+++ score/static_reflection_with_serialization/visitor/BUILD
@@ -42,7 +42,7 @@ test_suite(
         ":visitor_ut",
     ],
     visibility = [
-        # "@ddad@score_baselibs//score/mw/log/test:__pkg__",
+        # "@ddad@score_baselibs//score/mw/log/tests:__pkg__",
     ],
 )
 
@@ -56,8 +56,8 @@ cc_test(
     features = _compiler_warning_features,
     tags = ["unit"],
     visibility = [
-        #"@score_baselibs//score/mw/log/test:__pkg__",
-        # "@ddad@score_baselibs//score/mw/log/test:__pkg__",
+        #"@score_baselibs//score/mw/log/tests:__pkg__",
+        # "@ddad@score_baselibs//score/mw/log/tests:__pkg__",
     ],
     deps = [
         ":visitor",
